Create Custimation of AnKing MCAT deck stylization changes
I for the life of me did not at all one bit tolerate the fact that the standard deck was coded to be styled in what is essentially night mode and it took many many searches to figure out how to make this change. So I customized the day mode to be like typical anki with a white background and then changed the night mode too.